pyproject.toml
[tool.poetry]
name = "Ligh-tech-RND"
version = "0.1.0"
description = "test"
authors = ["Илья <rissol1986@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
gunicorn = "^22.0.0"
djangorestframework = "^3.15.2"
djoser = "^2.2.3"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
django-filter = "^24.3"
pytest = "^8.3.2"
pytest-django = "^4.8.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
ruff = "^0.4.4"
pre-commit = "^3.7.1"
testcontainers = "^4.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
